+++ fastapi-service/.python-version
@@ -1,0 +1,1 @@
+3.12
