+++ fastapi-service/src/app.py
@@ -1,0 +1,8 @@
+from fastapi import FastAPI
+
+from .cv_models.router import router as cv_models_router
+
+
+app = FastAPI()
+
+app.include_router(cv_models_router)
