+++ fastapi-service/src/cv_models/config.py
@@ -1,0 +1,8 @@
+from ..config import BaseConfig
+
+
+class CVModelGRPCConfig(BaseConfig):
+    grpc_host: str
+
+
+config = CVModelGRPCConfig()
