+++ fastapi-service/src/rq/dependencies.py
@@ -1,0 +1,5 @@
+from .worker_settings import q
+
+
+def get_rq_queue():
+    return q
