+++ grpc-cv-service/src/cv_models/core/Karla-VariableFont_wght.ttf
