--- pyproject.toml
@@ -1,28 +1,0 @@
-[tool.poetry]
-name = "money-counter-backend"
-version = "0.1.0"
-description = ""
-authors = ["Ivan Kirov"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.10"
-fastapi = {extras = ["all"], version = "^0.115.8"}
-pillow = "^11.1.0"
-sqlalchemy = "^2.0.38"
-ultralytics = "^8.3.74"
-onnxruntime = "^1.20.1"
-aiosqlite3 = "^0.3.0"
-aiosqlite = "^0.21.0"
-dramatiq = {extras = ["watch"], version = "^1.17.1"}
-
-[tool.poetry.group.dev.dependencies]
-mypy = "^1.15.0"
-
-[tool.mypy]
-strict = true
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
