--- src/database/config.py
@@ -1,9 +1,0 @@
-from config import BaseConfig
-
-
-class DBConfig(BaseConfig):
-
-    db_url: str
-
-
-config = DBConfig()
