--- src/roles/schemas.py
@@ -1,7 +1,0 @@
-from enum import StrEnum
-
-
-class Role(StrEnum):
-
-    CLIENT = 'client'
-    MANAGER = 'manager'
