--- src/users/dependencies.py
@@ -1,6 +1,0 @@
-from .services import UserService
-from .repositories import UserRepository
-
-
-def get_user_service() -> UserService:
-    return UserService(UserRepository)
