+++ telegram-bot/pyproject.toml
@@ -1,0 +1,11 @@
+[project]
+name = "telegram-bot"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "aiogram>=3.21.0",
+    "aiohttp>=3.12.15",
+    "pydantic-settings>=2.10.1",
+]
