+++ telegram-bot/src/commands/start.py
@@ -1,0 +1,10 @@
+from aiogram.filters import CommandStart
+from aiogram import Router
+from aiogram.types import Message
+
+
+router = Router()
+
+@router.message(CommandStart())
+async def start(message: Message):
+    pass
