+++ telegram-bot/src/dependencies.py
@@ -1,0 +1,6 @@
+from .services import CVModelsService
+from .repositories import FastAPIServiceRepository
+
+
+def get_cv_model_service() -> CVModelsService:
+    return CVModelsService(FastAPIServiceRepository)
